Improving hash coverage
