+++ Assets/LeapMotionModules/ElementRenderer/Scripts/LeapGuiElement.cs
@@ -137,7 +137,6 @@ public class LeapGuiElement : MonoBehaviour {
 
   void OnEnable() {
 #if UNITY_EDITOR
-    updateParentGui();
     if (Application.isPlaying) {
 #endif
       if (!IsAttachedToGui) {
@@ -150,7 +149,6 @@ public class LeapGuiElement : MonoBehaviour {
 
   void OnDisable() {
 #if UNITY_EDITOR
-    updateParentGui();
     if (Application.isPlaying) {
 #endif
       if (IsAttachedToGui) {
@@ -166,13 +164,6 @@ public class LeapGuiElement : MonoBehaviour {
     if (_pickingMesh != null && _pickingMesh.vertexCount != 0) {
       Gizmos.color = new Color(1, 0, 0, 0);
       Gizmos.DrawMesh(_pickingMesh);
-    }
-  }
-
-  private void updateParentGui() {
-    var gui = GetComponentInParent<LeapGui>();
-    if (gui != null) {
-      gui.ScheduleFullUpdate();
     }
   }
 #endif
