+++ Assets/LeapMotionModules/ElementRenderer/Scripts/Utility/HashUtil.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: e3cb8e0a268136d4db7ab5c1d672ccb7
+timeCreated: 1488908021
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
